GW-80: Updated GW-80
