+++ client/src/pages/contacts/components/PlugBox.jsx
@@ -7,7 +7,7 @@ export default function PlugBox() {
 		window.open(link, '_blank', 'noreferrer');
 	};
 	return (
-		<Box className='flex justify-between space-x-16 my-10'>
+		<Box className='flex justify-between space-x-10 lg:space-x-16 my-10'>
 			{PLUGS.map((plug) => (
 				<Box
 					className='bg-bgPrimary w-1/2 border-2 border-[#F2F0E3] rounded-2xl hover:cursor-pointer'
@@ -15,14 +15,14 @@ export default function PlugBox() {
 						handlePlugClick(plug.link);
 					}}
 				>
-					<Box className='p-8'>
+					<Box className='p-6 lg:p-8'>
 						<img src={plug.icon} alt='social-media-icon' />
 						<Box className='flex flex-col space-y-3 mt-10 lg:mt-14'>
 							<Typography variant='h5' color='primary'>
 								{plug.name}
 							</Typography>
-							<Typography variant='body2'>{plug.description}</Typography>
-							<Typography variant='body2' sx={{ textDecoration: 'underline', fontWeight: 'bold' }}>
+							<Typography variant='body1'>{plug.description}</Typography>
+							<Typography variant='body1' sx={{ textDecoration: 'underline', fontWeight: 'bold' }}>
 								{plug.handle}
 							</Typography>
 						</Box>
